# src/main/java/hei/school/agricole/dto/CollectivityInformation.java
package hei.school.agricole.dto;

public class CollectivityInformation {

    private String name;
    private String number;

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getNumber() {
        return number;
    }

    public void setNumber(String number) {
        this.number = number;
    }
}