# src/main/java/hei/school/agricole/dto/CreateCollectivityStructure.java
package hei.school.agricole.dto;

public class CreateCollectivityStructure {

    private String president;
    private String vicePresident;
    private String treasurer;
    private String secretary;

    public String getPresident() {
        return president;
    }

    public void setPresident(String president) {
        this.president = president;
    }

    public String getVicePresident() {
        return vicePresident;
    }

    public void setVicePresident(String vicePresident) {
        this.vicePresident = vicePresident;
    }

    public String getTreasurer() {
        return treasurer;
    }

    public void setTreasurer(String treasurer) {
        this.treasurer = treasurer;
    }

    public String getSecretary() {
        return secretary;
    }

    public void setSecretary(String secretary) {
        this.secretary = secretary;
    }
}